+++ lib/src/io/io.dart
@@ -1,7 +1,9 @@
 export 'asset_gen.dart';
-export 'commands.dart';
 export 'asset_gen.dart';
+export 'commands.dart';
 export 'dart_gen.dart';
 export 'env_parser.dart';
+export 'extract_strings.dart';
+export 'intl_dart_gen.dart';
 export 'sheet_parser.dart';
 export 'yaml_parser.dart';